Add spawn data for first map (needs virus balance)

+++ Assets/Scripts/Maps/Data/SpawnChances.cs
@@ -6,7 +6,7 @@ namespace FishingPrototype.Gameplay.Maps.Data
     [System.Serializable]
     public struct SpawnChances
     {
-        [HideInInspector] public List<float> chanceTypeList;
+        [HideInInspector] public List<int> chanceTypeList;
         [HideInInspector] public List<int> chanceAmountMinList;
         [HideInInspector] public List<int> chanceAmountMaxList;
     }
